refactor: `useDeleteToken` -> `useUser`로 훅이름 변경 #85
