+++ src/components/common/NavigateList.tsx
@@ -1,4 +1,4 @@
-import useDeleteToken from '@hooks/useDeleteToken';
+import useDeleteToken from '@hooks/useUser';
 import Link from 'next/link';
 
 type NavigateListProps = {
@@ -21,7 +21,7 @@ const NavigateList = ({ text, path }: NavigateListProps) => {
         className="flex h-[47px] items-center justify-center bg-white px-7"
         onClick={handleLogout}
       >
-        <span className="text-16 font-medium whitespace-nowrap hover:font-bold hover:text-brand">
+        <span className="whitespace-nowrap text-16 font-medium hover:font-bold hover:text-brand">
           {text}
         </span>
       </li>
